+++ tarantool/tlua/src/lib.rs
@@ -149,6 +149,7 @@ pub use rust_tables::{PushIterError, PushIterErrorOf, TableFromIter};
 pub use tuples::{AsTable, TuplePushError};
 pub use userdata::UserdataOnStack;
 pub use userdata::{push_some_userdata, push_userdata, read_userdata};
+pub use values::BytesInLua;
 pub use values::{False, Nil, Null, Strict, StringInLua, ToString, True, Typename};
 
 #[deprecated = "Use `CallError` instead"]