+++ tarantool/CHANGELOG.md
@@ -51,6 +51,8 @@
 - tlua::LuaFunction::load_file_contents function which allows to provide a
   filename for better lua error messages.
 - Clarification difference between `tlua::Nil` & `tlua::Null` in doc-comments.
+- tlua::BytesInLua helper struct for reading non-utf8 strings from lua without
+  extra allocations.
 
 ### Fixed